+++ mod-record-specifications-server/src/main/java/org/folio/rspec/controller/SpecificationStorageController.java
@@ -1,10 +1,13 @@
 package org.folio.rspec.controller;
 
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.folio.rspec.domain.dto.Family;
 import org.folio.rspec.domain.dto.FamilyProfile;
 import org.folio.rspec.domain.dto.IncludeParam;
 import org.folio.rspec.domain.dto.SpecificationDtoCollection;
+import org.folio.rspec.domain.dto.SpecificationRuleDtoCollection;
+import org.folio.rspec.domain.dto.ToggleSpecificationRuleDto;
 import org.folio.rspec.rest.resource.SpecificationStorageApi;
 import org.folio.rspec.service.SpecificationService;
 import org.springframework.http.HttpStatus;
@@ -23,10 +26,22 @@ public class SpecificationStorageController implements SpecificationStorageApi {
   public ResponseEntity<SpecificationDtoCollection> getSpecifications(Family family, FamilyProfile profile,
                                                                       IncludeParam include, Integer limit,
                                                                       Integer offset) {
-    if (include != null) {
+    if (IncludeParam.REQUIRED_FIELDS == include) {
       return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
     }
     var specifications = specificationService.findSpecifications(family, profile, include, limit, offset);
     return ResponseEntity.ok(specifications);
+  }
+
+  @Override
+  public ResponseEntity<SpecificationRuleDtoCollection> getSpecificationRules(UUID id) {
+    return ResponseEntity.ok(specificationService.findSpecificationRules(id));
+  }
+
+  @Override
+  public ResponseEntity<Void> toggleSpecificationRule(UUID specificationId, UUID ruleId,
+                                                      ToggleSpecificationRuleDto toggleSpecificationRuleDto) {
+    specificationService.toggleSpecificationRule(specificationId, ruleId, toggleSpecificationRuleDto);
+    return ResponseEntity.noContent().build();
   }
 }