+++ mod-record-specifications-server/src/main/java/org/folio/rspec/service/mapper/SpecificationRuleMapper.java
@@ -1,0 +1,22 @@
+package org.folio.rspec.service.mapper;
+
+import org.folio.rspec.domain.dto.SpecificationRuleDto;
+import org.folio.rspec.domain.entity.SpecificationRule;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.MappingConstants;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(unmappedTargetPolicy = ReportingPolicy.WARN, componentModel = MappingConstants.ComponentModel.SPRING,
+        uses = {MetadataMapper.class})
+public interface SpecificationRuleMapper {
+
+  @Mapping(target = "id", source = "rule.id")
+  @Mapping(target = "code", source = "rule.code")
+  @Mapping(target = "name", source = "rule.name")
+  @Mapping(target = "specificationId", source = "specification.id")
+  @Mapping(target = "description", source = "rule.description")
+  @Mapping(target = "metadata", source = "rule.metadata")
+  SpecificationRuleDto toDto(SpecificationRule specificationRule);
+
+}