+++ mod-record-specifications-server/src/test/java/org/folio/rspec/domain/repository/SpecificationRuleRepositoryIT.java
@@ -1,0 +1,47 @@
+package org.folio.rspec.domain.repository;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.support.TestConstants.BIBLIOGRAPHIC_SPECIFICATION_ID;
+import static org.folio.support.TestConstants.MISSING_FIELD_RULE_ID;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+
+import org.folio.rspec.domain.entity.SpecificationRuleId;
+import org.folio.spring.testing.extension.EnablePostgres;
+import org.folio.spring.testing.type.IntegrationTest;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+@DataJpaTest
+@EnablePostgres
+@IntegrationTest
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+class SpecificationRuleRepositoryIT {
+
+  @Autowired
+  private SpecificationRuleRepository specificationRuleRepository;
+
+  @Test
+  void findBySpecificationId() {
+    var specificationRuleList = specificationRuleRepository.findBySpecificationId(BIBLIOGRAPHIC_SPECIFICATION_ID);
+
+    assertThat(specificationRuleList)
+      .isNotEmpty()
+      .hasSize(12);
+  }
+
+  @Test
+  void updateEnabledBySpecificationRuleId() {
+    var specificationRuleId = new SpecificationRuleId(BIBLIOGRAPHIC_SPECIFICATION_ID, MISSING_FIELD_RULE_ID);
+    var updated = specificationRuleRepository.updateEnabledBySpecificationRuleId(false, specificationRuleId);
+
+    assertThat(updated).isEqualTo(1);
+
+    var updatedEntity = specificationRuleRepository.findById(specificationRuleId);
+    assertThat(updatedEntity)
+      .isNotEmpty()
+      .hasValueSatisfying(specificationRule -> assertFalse(specificationRule.isEnabled()));
+  }
+
+}